% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHOWLIU-package.R
\docType{package}
\name{CHOWLIU-package}
\alias{CHOWLIU}
\alias{CHOWLIU-package}
\title{CHOWLIU: Algorithme de CHOW-LIU}
\description{
Implementation de l'agorithme de CHOW-LIU de façon naive et en utilisant la methode de Prim.
}
\keyword{internal}
