% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual_information.R
\name{mutual_information}
\alias{mutual_information}
\title{Pour caculer l'information mutuelle d'un graphe, on applique la fonction ci-après à la matrice de variance covariance des variables (sommets du graphe)}
\usage{
mutual_information(cov_mat)
}
\arguments{
\item{cov_mat}{matrice de variance covariance}
}
\value{
Les sommets d'un graphe représentent les variables.
}
\description{
Pour caculer l'information mutuelle d'un graphe, on applique la fonction ci-après à la matrice de variance covariance des variables (sommets du graphe)
}
