% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Naive.R
\name{Naive_R}
\alias{Naive_R}
\title{Algorithme de Naive}
\usage{
Naive_R(weight_mat)
}
\arguments{
\item{weight_mat}{matrice d'adjacence (matrice des poids  qui lit  les différents sommets d'un graphe)}
}
\value{
l'arbre couvrant  de poids maximal
}
\description{
Algorithme de Naive
}
