% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prim_max.R
\name{Prim_max_R}
\alias{Prim_max_R}
\title{Algorithme de Prim}
\usage{
Prim_max_R(weight_mat)
}
\arguments{
\item{weight_mat}{matrice d'adjacence (matrice des poids  qui lit  les différents sommets d'un graphe)}
}
\value{
l'arbre couvrant  de poids maximal
}
\description{
Algorithme de Prim
}
